Linked List Traversal using Recursion.
Traversal means “visiting” or examining each node of the list. We start from the beginning and visit one node at a time until the end of the list (until the ‘next’ is NULL). As discussed in the previous post, we need the first element (head) to reach to any element of the list. So, we will do the traversal using the ‘head’ and print its element and then proceed to the next element of the list.

Thus, the steps for the traversal of the linked list are:

Check if the element is not NULL.
If it is not, then print its ‘data’.
Change the element to the element stored in the ‘next’.